Make compatible with old helixPropertiesAndOrigins
CTK uses python3.5 on Ubuntu 16.04LTS and douglaslab/cadnano2.5, a fork of turalaksel/cadnano2.5 as of March 10, 2018. Downstream commits to cadnano/cadnano2.5 add static type checking but break compaibility with python3.5, and we don't want to deal with upgrading to python3.6 on 16.04. Thus, this branch uses deprecated methods for accessing the model.